+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.2] - 2024-01-08
+
+### Changed
+- Default locale changed from 'en' to 'auto' - now detects browser language automatically
+- Added getBrowserLanguage() utility to detect user's browser language
+
+### Enhanced
+- Locale 'auto' setting will now use navigator.language to detect the user's preferred language
+- Falls back to 'en' if browser language cannot be detected
+
 ## [3.0.1] - 2024-01-08
 
 ### Changed